pyproject.toml
[project]
name = "pdf-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pandas>=2.3.1",
    "pdfminer-six[image]>=20250506",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
    "ruff>=0.12.7",
 "ty>=0.0.1a16",
]
